refactor: move filesystem bridge to platform-specific code
