+++ apps/desktop/src/main/services/index.ts
@@ -1,5 +1,4 @@
 // Barrel file for Node.js/Electron main process services
 // Export platform-specific service implementations here
 
-// Example exports (will be uncommented when implementations are added):
-// export { NodeFileSystemBridge } from './NodeFileSystemBridge';
+export { NodeFileSystemBridge } from "./NodeFileSystemBridge";
