+++ packages/shared/src/services/storage/index.ts
@@ -2,7 +2,6 @@ export * from "./errors";
 export * from "./TextEncoding";
 export * from "./WriteFileOptions";
 export * from "./FileSystemBridge";
-export * from "./NodeFileSystemBridge";
 export * from "./FileStorageService";
 export type { SecureStorageInterface } from "./SecureStorageInterface";
 export * from "./utils";
